2231 Enhance tensor transforms (#2966)
* [DLMED] enhance tensor transforms


* [DLMED] fix tests


* [DLMED] fix mypy

+++ tests/test_savitzky_golay_smooth.py
@@ -62,7 +62,7 @@ class TestSavitzkyGolaySmooth(unittest.TestCase):
     @parameterized.expand([TEST_CASE_SINGLE_VALUE, TEST_CASE_2D_AXIS_2, TEST_CASE_SINE_SMOOTH])
     def test_value(self, arguments, image, expected_data, atol):
         for p in TEST_NDARRAYS:
-            result = SavitzkyGolaySmooth(**arguments)(p(image))
+            result = SavitzkyGolaySmooth(**arguments)(p(image.astype(np.float32)))
             torch.testing.assert_allclose(result, p(expected_data.astype(np.float32)), rtol=1e-7, atol=atol)
 
 
@@ -70,7 +70,7 @@ class TestSavitzkyGolaySmoothREP(unittest.TestCase):
     @parameterized.expand([TEST_CASE_SINGLE_VALUE_REP])
     def test_value(self, arguments, image, expected_data, atol):
         for p in TEST_NDARRAYS:
-            result = SavitzkyGolaySmooth(**arguments)(p(image))
+            result = SavitzkyGolaySmooth(**arguments)(p(image.astype(np.float32)))
             torch.testing.assert_allclose(result, p(expected_data.astype(np.float32)), rtol=1e-7, atol=atol)
